Rework header and add brand
